+++ src/types.ts
@@ -9,10 +9,16 @@ export interface IInstance {
     readonly hostNport : string;
     readonly token : string;
     readonly org : string;
+    readonly orgID ?: string;
     readonly user : string;
     readonly pass : string;
     isActive : boolean;
     // Until there is a way to "migrate" this data, it may not
     // exist in the data store
     readonly disableTLS : boolean | undefined;
+}
+
+export interface IMigration {
+    readonly name : string;
+    readonly appliedOn : Date;
 }
