+++ templates/editConn.html
@@ -62,6 +62,7 @@
 		</div>
 
 		<input type="hidden" id="connID" value="{{id}}">
+		<input type="hidden" id="orgID" value="{{orgID}}">
 
 		<div class="actions">
 			<span class="teal button" id="testConn" tabindex="0">Test</span>
